assignment-2-finished
When I run the draw function drawing a set of sinc waves in e, trying
to draw as much as the example in d shown, the browser had difficulties
to execute the program. I guess it is because HTML5’s canvas occupy to
much CPU of the computer if I draw all the curves point-by-point. I
hope here is some other way to draw them.