ps-03_final
This project is done by myself and Lia Petronio we worked as a team
utilizing Lia’s code skills and my design skills. We intend to make
another view of the data and it is in progress.
In this folder is a pdf that is the mockup for the final goal.