finish one solution and get stuck for a long time.....
My DOM can not be removed every time I try to updated them. According
to what I’ve learned last semester, we should do “enter—exit—update” to
get data updated correctly, but that doesn’t work for this data. I
guess it is because something happens on “spreadsheet.js” to get the
data and refresh the data over and over….. I have no idea what to do
with this.